add 'Slugify' command to text formatting options
A 'Slugify' command was added to the list of text formatting commands in README.md. The 'Slugify' function transforms any given input text into a URL-friendly string by removing special characters and spaces and converting the text to lowercase. This was done to improve the functionality of text formatting and to ensure its relevance in creating SEO-friendly URLs or filenames. Included is an illustrative example showing the before and after effects of applying the 'Slugify' command on a string.